[spec] Move spec into spec/ folder.
This PR moves the spec file into a spec/ folder. A simple Makefile is
added into spec/ and a top level compile.sh script is added. The
spec is copied into the docs/ folder currently after being built.

The spec is updated to directly include the referenced svg/ images
and the paths are updated to point to `../src/static` as needed.