Rename src/ to explainer/
This PR changes the src/ directory to the explainer/ folder. The top level
README is moved into explainer/ and a new top level README is created. This
makes it a lot clearer the distinction between the spec and the explainer
documentation.

The src/meetings folder is moved to the top level meetings/ to disambiguate
the meeting minutes from the explainer content.

Various spec URLs have been updated to point to the explainer static files.