Reference Referer in id_token_endpoint fetch.
This PR adds the `Referer` to the `id_token_endpoint` requirements as
referenced in the provided example.

The `client_id` is already sent, but the `client_id` is considered
public information, providing the `Referer` provides the IDP with
another piece of information about the RP.

This maybe removed in the future depending on discussions, but this
makes the spec consistent with itself (and it's easier to make it
optional later then required).

Issue #114