Updates for CI action
This PR explicitly sets the TOOLCHAIN to bikeshed for the w3c/spec-prod
action.

The referenced mocks have been moved into the spec folder so they can
work with the build system. We previously used an override to allow
bikeshed to look in other folders but that doesn't seem to be possible
with the CI action. This makes it clearer which mocks are associated
with the spec and which are only in the explainer.

Remove explicit inheritance for `FederatedCredential` from the WebIDL
as the partial interface should not have inheritance specified.

Update `void` in IDL to `undefined`