Update the FedCM API.
This PR updates the FedCM API to be a closer match to the Credential
Management Level 1 API. The `get` method is changed to return the
`credential` instead of an `id_token`. A new `fetchTokens` method is
added in order to retrieve tokens. The `preventSilentAccess` is moved
behind an explicit `logout` API call. The IDP `logout` is renamed to
`idpLogout` to remove confusion.

Issue #145