Rename explainer to explorations.
The current explainer folder holds an exploration into the design space
that FedCM is considering. It takes about a lot of possible things that
could be done and ways to do that.

This is different from the explainer that will need to be produced which
explains the FedCM API specifically.

This PR moves the `explainer` folder to an `explorations` folder to make
room for the future explainer.