Update client metadata example to use absolute urls (#168)
* Make nonce optional in FederatedIdentityProvider

As suggested by some IDP, nonce should not be required. When RP doesn't specify a nonce, IDP will create one for them.

* Update client metadata example to use absolute urls

The client metadata endpoint is on idp.example so using relative urls could be confusing. This PR updates the example to use the absolute urls with the rp domain to reduce ambiguity.