Update and clarify the spec examples.
This PR starts updating the spec to clarify the return value of the
`get` request and make it more consistent. With this new change the
`get` request will never return a `null` value but will return a
FederatedCredential object with the correct consented space.

A followup PR will update the algorithms and other portions of the
spec to reflect this change.

Issue #145