Make "email" required in the accounts response
When a user is signed in to an IDP with multiple accounts and each of them uses the same name and avatar, the UI would be confusing without an extra bit (e.g. email) for them to differentiate.

This PR makes "email" required and we are open to adjust to relax it if there's need.