Reference Permissions Policy for Cross-Origin iframe Support (#233)
I think there's a misconception in the explainer that PP would only be useful for per-page control. It actually supports both iframe and header controls. Assuming that the header feature is not harmful to FedCM, PP is the recommended way of adding new `allow` features, makes it easier for implementers to support and more versatile for web developers to work with.