Major restructuring pass of non-core elements of the specification
This set of changes attempts to improve the clarity and structure of
the specification outside of the IDP API and Browser API sections.

This should address the following issues from Issue #416: 1-6, 10, 11, 14, 17, 18, 20, 21.

The Markdown files added here are drafts- converting Bikeshed to Markdown is non-trivial and
since they are targetting a return to the core spec it may not be worth the effort to make
a good looking version of the Markdown before creating a PR for them.