Adjust links to RFC9110 sections
The cross-reference database used in Bikeshed used to contain the `httpwg.org`
version of RFC9110. It now uses the `www.rfc-editor.org` version, consistent
with the URL used in the References section.

This update adjusts the links to RFC9110 sections accordingly.

(Note: the cross-reference database contains the actual section IDs, e.g.
`section-10.2.2`, not the section heading's IDs such as `name-location`)