Update links to RFC9110 (#520)
We rolled back on the change introduced in #452 and now use the more
readable `httpwg.org` version of RFC9110 to populate the cross-references
database. We also updated Specref to return the `httpwg.org` URL for that
spec.

This means that the cross-references database no longer has the section IDs
and rather has the field name. This update rolls back the changes in #452
accordingly.