Delete explainer.md
This explainer is out of date and per #622 we have been using the following resources. Removing the explainer to reduce confusion.

The spec is certainly up to date with Chrome's implementation, but it is probably harder to read than what we'd want to point developers too.

Chrome's developer documentation is also up to date, but we probably want something Chrome-agnostic.

MDN's developer documentation is chrome-agnostic, but not as up to date as Chrome's.